refractor: fix: refractor some minor code, and fix some codes issues
